# docs/api/api_swagger.yaml
basePath: /api
definitions:
  github_com_slxar_rustdesk-api-goravel_v3_http_request_api.AddressBookForm:
    properties:
      data:
        example: '{"tags":["tag1","tag2","tag3"],"peers":[{"id":"abc","username":"abv-l","hostname":"","platform":"Windows","alias":"","tags":["tag1","tag2"],"hash":"hash"}],"tag_colors":"{\"tag1\":4288585374,\"tag2\":4278238420,\"tag3\":4291681337}"}'
        type: string
    type: object
...
  github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  github_com_slxar_rustdesk-api-goravel_v3_http_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  model.AddressBookCollection:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - name
    type: object
  model.Tag:
    properties:
      collection:
        $ref: '#/definitions/model.AddressBookCollection'
      collection_id:
        type: integer
      color:
        description: color 是flutter的颜色值,从0x00000000 到 0xFFFFFFFF; 前两位表示透明度，后面6位表示颜色,
          可以转成rgba
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  model.TagList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: 接口
  title: 管理系统API
  version: "1.0"
paths:
...
  /ab:
    get:
      consumes:
      - application/json
      description: 地址列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 地址列表
      tags:
      - 地址
    post:
      consumes:
      - application/json
      description: 地址更新
      parameters:
      - description: 地址表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_request_api.AddressBookForm'
      produces:
      - application/json
      responses:
        "200":
          description: "null"
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 地址更新
      tags:
      - 地址
  /ab/peer/add/{guid}:
    delete:
      consumes:
      - application/json
      description: 删除地址
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除地址
      tags:
      - 地址[Personal]
    post:
      consumes:
      - application/json
      description: 添加地址
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 添加地址
      tags:
      - 地址[Personal]
  /ab/peer/update/{guid}:
    put:
      consumes:
      - application/json
      description: 更新地址
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新地址
      tags:
      - 地址[Personal]
  /ab/peers:
    get:
      consumes:
      - application/json
      description: 地址
      parameters:
      - description: 页码
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: guid
        in: query
        name: ab
        type: string
      - description: Device ID pattern (SQL LIKE wildcards)
        in: query
        name: id
        type: string
      - description: Alias pattern (SQL LIKE wildcards)
        in: query
        name: alias
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
      security:
      - BearerAuth: []
      summary: 地址列表
      tags:
      - 地址[Personal]
    post:
      consumes:
      - application/json
      description: 地址
      parameters:
      - description: 页码
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: guid
        in: query
        name: ab
        type: string
      - description: Device ID pattern (SQL LIKE wildcards)
        in: query
        name: id
        type: string
      - description: Alias pattern (SQL LIKE wildcards)
        in: query
        name: alias
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
      security:
      - BearerAuth: []
      summary: 地址列表
      tags:
      - 地址[Personal]
  /ab/personal:
    get:
      consumes:
      - application/json
      description: 个人地址
      parameters:
      - description: string valid
        in: body
        name: string
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
      security:
      - BearerAuth: []
      summary: 个人地址
      tags:
      - 地址[Personal]
    post:
      consumes:
      - application/json
      description: 个人地址
      parameters:
      - description: string valid
        in: body
        name: string
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
      security:
      - BearerAuth: []
      summary: 个人地址
      tags:
      - 地址[Personal]
  /ab/settings:
    post:
      consumes:
      - application/json
      description: 设置
      parameters:
      - description: string valid
        in: body
        name: string
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
      security:
      - BearerAuth: []
      summary: 设置
      tags:
      - 地址[Personal]
  /ab/shared/profiles:
    get:
      consumes:
      - application/json
      description: 共享
      parameters:
      - description: Exact address-book name
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
      security:
      - BearerAuth: []
      summary: 共享地址簿
      tags:
      - 地址[Personal]
    post:
      consumes:
      - application/json
      description: 共享
      parameters:
      - description: Exact address-book name
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.Response'
      security:
      - BearerAuth: []
      summary: 共享地址簿
      tags:
      - 地址[Personal]
  /ab/tag/{guid}:
    delete:
      consumes:
      - application/json
      description: 标签
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 标签删除
      tags:
      - 地址[Personal]
  /ab/tag/add/{guid}:
    post:
      consumes:
      - application/json
      description: 标签
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 标签添加
      tags:
      - 地址[Personal]
  /ab/tag/rename/{guid}:
    put:
      consumes:
      - application/json
      description: 标签
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 标签重命名
      tags:
      - 地址[Personal]
  /ab/tag/update/{guid}:
    put:
      consumes:
      - application/json
      description: 标签
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 标签修改颜色
      tags:
      - 地址[Personal]
  /ab/tags/{guid}:
    get:
      consumes:
      - application/json
      description: 标签
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TagList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 标签
      tags:
      - 地址[Personal]
    post:
      consumes:
      - application/json
      description: 标签
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TagList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_slxar_rustdesk-api-goravel_v3_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 标签
      tags:
      - 地址[Personal]
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"